Moved Chat to separate file
